+++ backend/lib/backend/application.ex
@@ -33,7 +33,8 @@ defmodule Backend.Application do
       ),
       Supervisor.child_spec({Task, fn -> HTTPoison.start() end}, id: :start_httpoison),
       Backend.Scheduler,
-      Backend.Elasticsearch.Cluster
+      Backend.Elasticsearch.Cluster,
+      Graph.Cache
     ]
 
     children =