+++ backend/lib/backend_web/controllers/instance_controller.ex
@@ -2,17 +2,17 @@ defmodule BackendWeb.InstanceController do
   use BackendWeb, :controller
 
   import Backend.Util
-  alias Backend.Api
+  alias Graph.Cache
 
   action_fallback(BackendWeb.FallbackController)
 
   def show(conn, %{"id" => domain}) do
-    instance = Api.get_instance_with_peers(domain)
+    instance = Cache.get_instance_with_peers(domain)
 
     if instance == nil or instance.opt_out == true do
       send_resp(conn, 404, "Not found")
     else
-      last_crawl = get_last_crawl(domain)
+      last_crawl = Cache.get_last_crawl(domain)
       render(conn, "show.json", instance: instance, crawl: last_crawl)
     end
   end