+++ backend/lib/backend/util.ex
@@ -202,4 +202,22 @@ defmodule Backend.Util do
       {:error, error} -> raise error
     end
   end
+
+  @doc """
+  Strips `prefix` from `string`. If it doesn't start with that prefix, just returns the string.
+  """
+  @spec strip_prefix(String.t(), String.t()) :: String.t()
+  def strip_prefix(string, prefix) do
+    if String.starts_with?(string, prefix) do
+      prefix_length = String.length(prefix)
+      String.slice(string, prefix_length..-1)
+    else
+      string
+    end
+  end
+
+  @spec is_valid_domain?(String.t()) :: boolean
+  def is_valid_domain?(domain) do
+    Regex.match?(~r/[\w.-_]+/, domain)
+  end
 end