+++ backend/lib/backend_web/endpoint.ex
@@ -46,7 +46,11 @@ defmodule BackendWeb.Endpoint do
   )
 
   plug(Corsica,
-    origins: ["http://localhost:3000", ~r{^https?://(.*\.?)fediverse\.space$}],
+    origins: [
+      "http://localhost:3000",
+      ~r{^https?://(.*\.?)fediverse\.space$},
+      ~r{^https?://(.*\.?)fediverse-space\.netlify.com$}
+    ],
     allow_headers: ["content-type", "token"]
   )