+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add caching to graph + instance endpoints to better handle traffic spikes.
+
 ### Changed
 
 ### Deprecated