+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added ON DELETE to `most_recent_crawl` table, such that it can handle previously-crawled but now-dead instances.
 - You can now login to the admin view by clicking, not just by pressing enter.
+- Add handling for weirdly-formatted Friendica peers
 
 ## [2.7.0 - 2018-08-18]